lstm_torch_tokenizer/requirements.txt
numba==0.51.2
tensorflow==2.3.0
seaborn==0.11.0
pandas==1.1.3
torchmetrics==0.2.0
lightning_bolts==0.3.3
matplotlib==3.3.2
tqdm==4.50.2
torch==1.7.1
numpy==1.18.5
PyYAML==5.4.1
rich==10.2.2
scikit_learn==0.24.2
